8261447: MethodInvocationCounters frequently run into overflow Reviewed-by: mdoerr, thartmann, kvn, iveresov

+++ src/hotspot/share/runtime/sharedRuntime.hpp
@@ -74,7 +74,7 @@ class SharedRuntime: AllStatic {
 
 #ifndef PRODUCT
   // Counters
-  static int     _nof_megamorphic_calls;         // total # of megamorphic calls (through vtable)
+  static int64_t _nof_megamorphic_calls;         // total # of megamorphic calls (through vtable)
 #endif // !PRODUCT
 
  private:
@@ -568,16 +568,16 @@ class SharedRuntime: AllStatic {
 
   // Statistics code
   // stats for "normal" compiled calls (non-interface)
-  static int     _nof_normal_calls;              // total # of calls
-  static int     _nof_optimized_calls;           // total # of statically-bound calls
-  static int     _nof_inlined_calls;             // total # of inlined normal calls
-  static int     _nof_static_calls;              // total # of calls to static methods or super methods (invokespecial)
-  static int     _nof_inlined_static_calls;      // total # of inlined static calls
+  static int64_t _nof_normal_calls;               // total # of calls
+  static int64_t _nof_optimized_calls;            // total # of statically-bound calls
+  static int64_t _nof_inlined_calls;              // total # of inlined normal calls
+  static int64_t _nof_static_calls;               // total # of calls to static methods or super methods (invokespecial)
+  static int64_t _nof_inlined_static_calls;       // total # of inlined static calls
   // stats for compiled interface calls
-  static int     _nof_interface_calls;           // total # of compiled calls
-  static int     _nof_optimized_interface_calls; // total # of statically-bound interface calls
-  static int     _nof_inlined_interface_calls;   // total # of inlined interface calls
-  static int     _nof_megamorphic_interface_calls;// total # of megamorphic interface calls
+  static int64_t _nof_interface_calls;            // total # of compiled calls
+  static int64_t _nof_optimized_interface_calls;  // total # of statically-bound interface calls
+  static int64_t _nof_inlined_interface_calls;    // total # of inlined interface calls
+  static int64_t _nof_megamorphic_interface_calls;// total # of megamorphic interface calls
   // stats for runtime exceptions
   static int     _nof_removable_exceptions;      // total # of exceptions that could be replaced by branches due to inlining
 
@@ -591,7 +591,7 @@ class SharedRuntime: AllStatic {
   static address nof_optimized_interface_calls_addr()   { return (address)&_nof_optimized_interface_calls; }
   static address nof_inlined_interface_calls_addr()     { return (address)&_nof_inlined_interface_calls; }
   static address nof_megamorphic_interface_calls_addr() { return (address)&_nof_megamorphic_interface_calls; }
-  static void print_call_statistics(int comp_total);
+  static void print_call_statistics(uint64_t comp_total);
   static void print_statistics();
   static void print_ic_miss_histogram();
